Fix SVG rendering issues in About Page
This solution will check if the app is in mobile every re render. This means that it will not work if you simply switch from desktop to mobile in the inspect element, you must switch it first and then reload to cause the re-render. This won't be an issue for mobile devices, because they will already be on mobile, and the initial render will indicate to the state that it's a mobile device.